# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/flow/apps/{app_id}": {
            "get": {
                "description": "查询租户流量",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据接口"
                ],
                "summary": "查询租户流量接口",
                "operationId": "/flow/apps/{app_id}",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app id",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.Flow"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/flow/services/{service_id}": {
            "get": {
                "description": "查询服务流量",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据接口"
                ],
                "summary": "查询服务流量接口",
                "operationId": "/flow/services/{service_id}",
                "parameters": [
                    {
                        "type": "string",
                        "description": "service id",
                        "name": "service_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.Flow"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/services": {
            "get": {
                "description": "查询服务列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务接口"
                ],
                "summary": "查询服务列表接口",
                "operationId": "/services",
                "parameters": [
                    {
                        "type": "string",
                        "description": "keyword",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page index",
                        "name": "page_index",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ListServicesRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/tokens/jwt": {
            "post": {
                "description": "生成认证令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证接口"
                ],
                "summary": "生成认证令牌接口",
                "operationId": "/tokens/jwt",
                "parameters": [
                    {
                        "description": "generate jwt request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.JwtReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.JwtRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/admin": {
            "get": {
                "description": "获取用户的信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "查询用户信息接口",
                "operationId": "/users/admin",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UserInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "patch": {
                "description": "修改用户的密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "修改用户密码接口",
                "operationId": "/users/admin",
                "parameters": [
                    {
                        "description": "change user password request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChangeUserPwdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common_middleware.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errmsg": {
                    "type": "string"
                },
                "errno": {
                    "type": "integer"
                },
                "stack": {
                    "type": "object"
                },
                "trace_id": {
                    "type": "object"
                }
            }
        },
        "dto.ChangeUserPwdReq": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
...
        "dto.Flow": {
            "type": "object",
            "properties": {
                "today_flow": {
                    "description": "今日流量列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "yesterday_flow": {
                    "description": "昨日流量列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "dto.JwtReq": {
            "type": "object",
            "properties": {
                "permission": {
                    "description": "读写权限",
                    "type": "string"
                },
                "type": {
                    "description": "令牌类型",
                    "type": "string"
                }
            }
        },
        "dto.JwtRes": {
            "type": "object",
            "properties": {
                "expire_at": {
                    "description": "失效时间",
                    "type": "integer"
                },
                "permission": {
                    "description": "读写权限",
                    "type": "string"
                },
                "token": {
                    "description": "令牌内容",
                    "type": "string"
                },
                "type": {
                    "description": "令牌类型",
                    "type": "string"
                }
            }
        },
...
        "dto.ListServiceItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "real_qpd": {
                    "type": "integer"
                },
                "real_qps": {
                    "type": "integer"
                },
                "service_addr": {
                    "type": "string"
                },
                "service_desc": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "service_type": {
                    "type": "integer"
                },
                "total_node": {
                    "type": "integer"
                }
            }
        },
        "dto.ListServicesRes": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "服务列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ListServiceItem"
                    }
                },
                "total": {
                    "description": "共计条数",
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "dto.UserInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "intro": {
                    "type": "string"
                },
                "login_at": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}